requirements.txt
google-cloud-aiplatform==1.53.0
fastapi==0.110.2
uvicorn==0.29.0
pandas==1.5.3
requests==2.31.0
Pillow==10.3.0


